requirements.txt
Flask
flask-cors
pandas
numpy
scikit-learn
joblib
pymongo
python-dotenv
gunicorn
requests